rename metanorma-acme to metanorma-generic: https://github.com/metanorma/metanorma-generic/issues/17
