+++ lib/isodoc/vsd/html_convert.rb
@@ -1,5 +1,5 @@
 require "isodoc"
-require "isodoc/acme/html_convert"
+require "isodoc/generic/html_convert"
 require "isodoc/vsd/metadata"
 
 module IsoDoc
@@ -7,7 +7,7 @@ module IsoDoc
     # A {Converter} implementation that generates HTML output, and a document
     # schema encapsulation of the document for validation
     #
-    class HtmlConvert < IsoDoc::Acme::HtmlConvert
+    class HtmlConvert < IsoDoc::Generic::HtmlConvert
       def configuration
         Metanorma::Vsd.configuration
       end
