+++ lib/isodoc/vsd/metadata.rb
@@ -2,7 +2,7 @@ require "isodoc"
 
 module IsoDoc
   module Vsd
-    class Metadata < IsoDoc::Acme::Metadata
+    class Metadata < IsoDoc::Generic::Metadata
       def configuration
         Metanorma::Vsd.configuration
       end
