+++ metanorma-vsd.gemspec
@@ -29,7 +29,7 @@ Gem::Specification.new do |spec|
   spec.add_dependency "htmlentities", "~> 4.3.4"
   #spec.add_dependency "nokogiri"
   spec.add_dependency "metanorma-standoc", "~> 1.3.0"
-  spec.add_dependency "metanorma-acme", "~> 1.4.0"
+  spec.add_dependency "metanorma-generic", "~> 1.4.0"
   spec.add_dependency "isodoc", "~> 1.0.0"
 
   spec.add_development_dependency "byebug", "~> 9.1"
